refeact(applicant) : 도메인 별로 서비스 로직 나누기
